requirements.txt
# Core AI/ML Libraries
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.40
openai>=1.12.0

# Vector Database
chromadb>=0.4.22

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.31.0

# CLI Framework
click>=8.1.7

# Data Processing
pandas>=2.1.4
beautifulsoup4>=4.12.2
requests>=2.31.0
markdown>=3.5.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
rich>=13.7.0

# Optional: For better embeddings
sentence-transformers>=2.3.1

